govihub-api/requirements.txt
fastapi>=0.110.0,<1.0
uvicorn[standard]>=0.27.0
gunicorn>=21.2.0
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.27.0
redis>=5.0.0
celery>=5.3.0
Pillow>=10.2.0
boto3>=1.34.0
python-multipart>=0.0.6
structlog>=24.1.0
orjson>=3.9.0
geoalchemy2>=0.14.0
pgvector>=0.2.0
sentence-transformers>=2.3.0
torch>=2.1.0
torchvision>=0.16.0
scikit-learn>=1.4.0
numpy>=1.26.0
firebase-admin>=6.3.0
twilio>=8.10.0
apscheduler>=3.10.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-httpx>=0.30.0
aiosqlite>=0.19.0
factory-boy>=3.3.0
authlib>=1.3.0
itsdangerous>=2.1.0
